+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Add client identifier to payload
 * Add rails 3.0 or higher requirement to gemspec
+* Add rake task to verify connectivity